Add assertEqualsShouldFailWith function to specHelper
Refactor the assertions spec and use a new function from the specHelper, 
assertEqualsShouldFailWith.

This function should test specs that are designed to throw errors, and 
pass them when the correct errors thrown. If the spec does not create an 
error, or the error is not correct, the spec should fail.