Add array elements assertion
The assertEquals function should go through each element in the array 
and check it matches the element in the same position of the passed 
actual.

If the element does not match the index position of the first non-match 
is returned in the error message, along with the expected and actual 
values.

The equality test between elements will only work for primitive types.