Add multidimensional array assertion
To assert whether 2 arrays that have elements that are arrays themselves 
are equal, the program will need to recursive loop through the arrays, 
calling assertEquals on each element rather than simple doing a 
primitive type equality check.

So that we can also track which element is currently being tested and 
potentially fails, an array stores the index on which we are recursively 
calling assertEquals, popping the element if it returns true.

To facilitate the recursive call and tracker, a new deep equality 
function is required that can accept a tracker array, so that a new one 
is not initialised everytime the recursive call is made.