Add object key exists assertion
To check if a key exists in both expected and actual values a switch 
statement can initially be used within the assertDeepEquality function 
to first identify the object type.

This can call assertObjectEquality, which converts the object into an 
array that can be iterated through. For each key expected check if it 
exists in the actual object.