Simplify assertion failure tracking
A message and stack array is currently being passed between functions 
that test for equality. Update function arguements and parameters so 
that a tracker object is passed instead.

The tracker object can hold both the stack and message related to any 
failures that occur.

This will also make it simpler to compose the exception messages and 
reduce ducplication of how it is presented.