Add object key values assertion
For each key within an object check both the expected and actual values 
are equal by passing them to assertDeepEquality, creating a recursive 
function.

This will mean any key values that are not primitives are correctly 
checked for equaility as their type will be evaluated and they will be 
routed through the switch statement.